minor fixes

+++ src/admin/module/config_upload.php
@@ -107,9 +107,9 @@ function admin_config_upload($engine, $module)
 
 
 		$config['upload']					= (string) $upload;
-		$config['upload_max_size']			= (int) ($_POST['upload_max_size'] * $binary_factor[$_POST['upload_max_size_factor']]);
-		$config['upload_quota']				= (int) ($_POST['upload_quota'] * $binary_factor[$_POST['upload_quota_factor']]);
-		$config['upload_quota_per_user']	= (int) $_POST['upload_quota_per_user'] * $binary_factor[$_POST['upload_quota_per_user_factor']];
+		$config['upload_max_size']			= (int) ($_POST['upload_max_size']			* $binary_factor[(int) $_POST['upload_max_size_factor']]);
+		$config['upload_quota']				= (int) ($_POST['upload_quota']				* $binary_factor[(int) $_POST['upload_quota_factor']]);
+		$config['upload_quota_per_user']	= (int) ($_POST['upload_quota_per_user']	* $binary_factor[(int) $_POST['upload_quota_per_user_factor']]);
 		$config['upload_translit']			= (int) $_POST['upload_translit'];
 		$config['upload_images_only']		= (int) ($_POST['upload_images_only'] ?? 0);
 		$config['upload_allowed_exts']		= (string) $allowed_exts;
